client_download_heightmap initial
