+++ crates/bounded_planet_core/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2018"
 bevy = "0.2.1"
 bevy_rapier3d = "0.2.0"
 tokio = { version = "0.2.22", features = ["full"] }
-quinn = "0.6.1"
 structopt = "0.3.17"
 tracing = "0.1"
 tracing-subscriber = "0.2"
@@ -23,10 +22,10 @@ deflate = "0.8.6"
 futures-util = "0.3.5"
 uuid = { version = "0.8", features = ["v4"] }
 itertools = "0.9.0"
-
 anyhow = "1.0.32"
 thiserror = "1.0.20"
 
+quinn = "0.6.1"
 # rustls isn't directly needed, it's a dependency of `quinn`. The `dangerous_configuration` feature is required to bypass security in the networking.
 # Make sure to recheck this versaion number when updating quinn!
 rustls = { version = "0.17.0", features = [ "dangerous_configuration" ] }
