+++ crates/bounded_planet_core/src/land/mod.rs
@@ -2,4 +2,8 @@ mod heightmap;
 pub use heightmap::TextureHeightmap;
 
 mod mesh;
+pub use mesh::texture_to_mesh_data;
 pub use mesh::texture_to_mesh;
+pub use mesh::MeshData;
+
+pub mod systems;
